Moves uniqueness query logic into correct class
DHIS2-6374
