+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/trackedentity/TrackedEntityAttributeStore.java
@@ -31,6 +31,7 @@ package org.hisp.dhis.trackedentity;
 import org.hisp.dhis.common.IdentifiableObjectStore;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * @author Abyot Asalefew Gizaw
@@ -57,4 +58,12 @@ public interface TrackedEntityAttributeStore
      */
     List<TrackedEntityAttribute> getDisplayInListNoProgram();
 
+    /**
+     * Check whether there already exists a TrackedEntityInstance with given unique attribute value. If yes, return
+     * Optional containing UID of it. Otherwise, return empty Optional.
+     *
+     * @param params Query params. Contains value of unique attribute that should be checked.
+     * @return Optional of TrackedEntityInstance UID or empty Optional.
+     */
+    Optional<String> getTrackedEntityInstanceUidWithUniqueAttributeValue( TrackedEntityInstanceQueryParams params );
 }
