+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/trackedentity/TrackedEntityInstanceStore.java
@@ -34,7 +34,6 @@ import org.hisp.dhis.user.User;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 /**
  * @author Abyot Asalefew Gizaw
@@ -77,13 +76,4 @@ public interface TrackedEntityInstanceStore
     void updateTrackedEntityInstancesSyncTimestamp( List<String> trackedEntityInstanceUIDs, Date lastSynchronized );
 
     List<TrackedEntityInstance> getTrackedEntityInstancesByUid( List<String> uids, User user );
-
-    /**
-     * Check whether there already exists a TrackedEntityInstance with given unique attribute value. If yes, return
-     * Optional containing UID of it. Otherwise, return empty Optional.
-     *
-     * @param params Query params. Contains value of unique attribute that should be checked.
-     * @return Optional of TrackedEntityInstance UID or empty Optional.
-     */
-    Optional<String> getTrackedEntityInstanceUidWithUniqueAttributeValue( TrackedEntityInstanceQueryParams params );
 }
